+++ packages/_private/vue-demo/src/AllComponentParts/index.ts
@@ -25,6 +25,7 @@ import DemoPagination from "./DemoPagination.vue";
 import DemoPrioNav from "./DemoPrioNav.vue";
 import DemoProgressBar from "./DemoProgressBar.vue";
 import DemoProgressRing from "./DemoProgressRing.vue";
+import DemoRadioButton from "./DemoRadioButton.vue";
 import DemoRadioGroup from "./DemoRadioGroup.vue";
 import DemoRange from "./DemoRange.vue";
 import DemoSelect from "./DemoSelect.vue";
@@ -65,6 +66,7 @@ export {
   DemoPrioNav,
   DemoProgressBar,
   DemoProgressRing,
+  DemoRadioButton,
   DemoRadioGroup,
   DemoRange,
   DemoSelect,