+++ packages/components/code-quality/html-validate/scripts/synergy-element-rules.js
@@ -353,6 +353,16 @@ const SynRadio = {
 /**
  * @type {import('html-validate').MetaElement}
  */
+const SynRadioButton = {
+  focusable: true,
+  formAssociated,
+  interactive: true,
+  phrasing: true,
+};
+
+/**
+ * @type {import('html-validate').MetaElement}
+ */
 const SynRangeTick = {
   flow: true,
   phrasing: true,
@@ -525,6 +535,7 @@ export const rules = {
   SynProgressBar,
   SynProgressRing,
   SynRadio,
+  SynRadioButton,
   SynRadioGroup,
   SynRange,
   SynRangeTick,