+++ packages/docs/src/decorators/index.ts
@@ -1,0 +1,5 @@
+export * from './FormDecorator.js';
+export * from './PaddingDecorator.js';
+export * from './StopAnimation.js';
+export * from './ThemeNotSupportedHint.js';
+export * from './ThemeSwitchIcons.js';